[ci skip] WIP

+++ pykeops/examples/pytorch/plot_generic_syntax_pytorch.py
@@ -66,7 +66,7 @@ variables = [
 # i.e. on the axis ``1`` of the kernel matrix.
 # The output c is an :math:`x`-variable indexed by :math:`i`.
 
-my_routine = Genred(formula, variables, reduction_op="Sum", axis=1, dtype=dtype)
+my_routine = Genred(formula, variables, reduction_op="Sum", axis=1)
 c = my_routine(x, y, a, p)
